feat: add state
