+++ src/index.html
@@ -20,7 +20,6 @@
               Controls
             </div>
             <div class="goods">
-              Goods
             </div>
           </div>
         </main>
