+++ src/scripts/state.ts
@@ -1,0 +1,25 @@
+interface Good {
+  id: number;
+  name: string;
+  year: number;
+}
+
+const goods: Good[] = [
+  {
+    id: 1,
+    name: 'Jordan 1',
+    year: 2015
+  },
+  {
+    id: 2,
+    name: 'Jordan 2',
+    year: 2012
+  },
+  {
+    id: 3,
+    name: 'Jordan 3',
+    year: 2011
+  }
+]
+
+export default goods;
